move twitter/ and its tests into cli/commands/twitter/ for self-contained CLI module
