+++ assistant/src/cli/commands/twitter/client.ts
@@ -4,7 +4,7 @@
  * go through the browser's authenticated session.
  */
 
-import { ProviderError } from "../util/errors.js";
+import { ProviderError } from "../../../util/errors.js";
 import { loadSession, type TwitterSession } from "./session.js";
 
 const CDP_BASE = "http://localhost:9222";
