+++ assistant/src/cli/commands/twitter/index.ts
@@ -9,7 +9,7 @@ import * as net from "node:net";
 
 import { Command } from "commander";
 
-import { createMessageParser, serialize } from "../../daemon/ipc-protocol.js";
+import { createMessageParser, serialize } from "../../../daemon/ipc-protocol.js";
 import {
   getBookmarks,
   getFollowers,
@@ -23,15 +23,15 @@ import {
   getUserTweets,
   searchTweets,
   SessionExpiredError,
-} from "../../twitter/client.js";
-import type { TwitterStrategy } from "../../twitter/router.js";
-import { routedPostTweet } from "../../twitter/router.js";
+} from "./client.js";
+import type { TwitterStrategy } from "./router.js";
+import { routedPostTweet } from "./router.js";
 import {
   clearSession,
   importFromRecording,
   loadSession,
-} from "../../twitter/session.js";
-import { getSocketPath, readSessionToken } from "../../util/platform.js";
+} from "./session.js";
+import { getSocketPath, readSessionToken } from "../../../util/platform.js";
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -309,7 +309,7 @@ Examples:
           {
             type: "twitter_integration_config",
             action: "get",
-          } as import("../../daemon/ipc-protocol.js").ClientMessage,
+          } as import("../../../daemon/ipc-protocol.js").ClientMessage,
           "twitter_integration_config_response",
         );
         const r = daemonResponse as Record<string, unknown>;
@@ -375,7 +375,7 @@ Examples:
           {
             type: "twitter_integration_config",
             action: "get_strategy",
-          } as import("../../daemon/ipc-protocol.js").ClientMessage,
+          } as import("../../../daemon/ipc-protocol.js").ClientMessage,
           "twitter_integration_config_response",
         );
         const r = daemonResponse as Record<string, unknown>;
@@ -412,7 +412,7 @@ Examples:
             type: "twitter_integration_config",
             action: "set_strategy",
             strategy: value,
-          } as import("../../daemon/ipc-protocol.js").ClientMessage,
+          } as import("../../../daemon/ipc-protocol.js").ClientMessage,
           "twitter_integration_config_response",
         );
         const r = daemonResponse as Record<string, unknown>;
@@ -879,7 +879,7 @@ Examples:
 // ---------------------------------------------------------------------------
 
 function sendDaemonMessage(
-  message: import("../../daemon/ipc-protocol.js").ClientMessage,
+  message: import("../../../daemon/ipc-protocol.js").ClientMessage,
   expectedResponseType: string,
 ): Promise<Record<string, unknown>> {
   return new Promise((resolve, reject) => {
@@ -960,7 +960,7 @@ function sendDaemonMessage(
           serialize({
             type: "auth",
             token: sessionToken,
-          } as unknown as import("../../daemon/ipc-protocol.js").ClientMessage),
+          } as unknown as import("../../../daemon/ipc-protocol.js").ClientMessage),
         );
       } else {
         sendPayload();
@@ -976,7 +976,7 @@ function sendDaemonMessage(
 import {
   ensureChromeWithCdp,
   minimizeChromeWindow,
-} from "../../tools/browser/chrome-cdp.js";
+} from "../../../tools/browser/chrome-cdp.js";
 
 // ---------------------------------------------------------------------------
 // Ride Shotgun learn session helper
@@ -1052,7 +1052,7 @@ async function startLearnSession(
           intervalSeconds: 5,
           mode: "learn",
           targetDomain: "x.com",
-        } as unknown as import("../../daemon/ipc-protocol.js").ClientMessage),
+        } as unknown as import("../../../daemon/ipc-protocol.js").ClientMessage),
       );
     };
 
@@ -1101,7 +1101,7 @@ async function startLearnSession(
           serialize({
             type: "auth",
             token: sessionToken,
-          } as unknown as import("../../daemon/ipc-protocol.js").ClientMessage),
+          } as unknown as import("../../../daemon/ipc-protocol.js").ClientMessage),
         );
       } else {
         sendStartCommand();
